implement ListCreateAPIView & RetrieveUpdateDestroyAPIView

+++ store/views.py
@@ -2,6 +2,9 @@ from django.db.models import Count
 from django.shortcuts import get_object_or_404
 from rest_framework import status
 from rest_framework.decorators import api_view
+from rest_framework.generics import (ListCreateAPIView,
+                                     RetrieveUpdateDestroyAPIView)
+from rest_framework.mixins import CreateModelMixin, ListModelMixin
 from rest_framework.response import Response
 from rest_framework.views import APIView
 
@@ -11,41 +14,27 @@ from .serializers import CollectionSerializer, ProductsSerializer
 # Create your views here.
 
 
-class ProductList(APIView):
+class ProductList(ListCreateAPIView):
+    queryset = Product.objects.select_related("collection").all()
+    serializer_class = ProductsSerializer
 
-    def get(self, request):
-        queryset = Product.objects.select_related("collection").all()
-        serializer = ProductsSerializer(
-            queryset, many=True, context={"request": request}
-        )
-        return Response(serializer.data)
+    def perform_create(self, serializer):
+        return super().perform_create(serializer)
 
-    def post(self, request):
-        serializer = ProductsSerializer(data=request.data)
-        serializer.is_valid(raise_exception=True)
-        serializer.save()
-        return Response(serializer.data, status=status.HTTP_201_CREATED)
+    def get_serializer_context(self):
+        return {"request": self.request}
 
 
-class ProductDetail(APIView):
-    def get(self, request, pk):
-        product = get_object_or_404(Product, pk=pk)
-        serializer = ProductsSerializer(product)
-        return Response(serializer.data)
-
-    def put(self, request, pk):
-        product = get_object_or_404(Product, pk=pk)
-        serializer = ProductsSerializer(product, data=request.data)
-        serializer.is_valid(raise_exception=True)
-        serializer.save()
-        return Response(serializer.data, status=status.HTTP_200_OK)
+class ProductDetail(RetrieveUpdateDestroyAPIView):
+    queryset = Product.objects.select_related("collection").all()
+    serializer_class = ProductsSerializer
 
     def delete(self, request, pk):
-        product = get_object_or_404(Product, pk=pk)
+        product = get_object_or_404(Product, id=pk)
         if product.orderitems.count() > 0:
             return Response(
                 {
-                    "error": "Product cannot be deleted because it is associated with an order items"
+                    "error": "Product cannot be delete because order items associates with this"
                 },
                 status=status.HTTP_405_METHOD_NOT_ALLOWED,
             )
@@ -53,37 +42,23 @@ class ProductDetail(APIView):
         return Response(status=status.HTTP_204_NO_CONTENT)
 
 
-class CollectionList(APIView):
-    def get(self, request):
-        quetyset = (
-            Collection.objects.
-            annotate(products_count=Count("products")).all()
-        )
-        serializer = CollectionSerializer(quetyset, many=True)
-        return Response(serializer.data, status=status.HTTP_200_OK)
-
-    def post(self, request):
-        serializer = CollectionSerializer(data=request.data)
-        serializer.is_valid(raise_exception=True)
-        serializer.save()
-        return Response(serializer.data, status=status.HTTP_201_CREATED)
+class CollectionCreateList(ListCreateAPIView):
+    queryset = Collection.objects.annotate(products_count=Count("products")).all()
+    serializer_class = CollectionSerializer
 
 
-class CollectionDetail(APIView):
-    def get(self, request, pk):
-        collection = get_object_or_404(
-            Collection.objects.annotate(products_count=Count("products")), pk=pk)
-        serializer = CollectionSerializer(collection)
-        return Response(serializer.data)
-
-    def put(self, request, pk):
-        collection = get_object_or_404(Collection, pk=pk)
-        serializer = CollectionSerializer(collection, data=request.data)
-        serializer.is_valid(raise_exception=True)
-        serializer.save()
-        return Response(serializer.data, status=status.HTTP_200_OK)
+class CollectionRetriveUpdateDestroy(RetrieveUpdateDestroyAPIView):
+    queryset = Collection.objects.annotate(products_count=Count("products")).all()
+    serializer_class = CollectionSerializer
 
     def delete(self, request, pk):
-        collection = get_object_or_404(Collection, pk=pk)
+        collection = get_object_or_404(Collection, id=pk)
+        if collection.products.count() > 0:
+            return Response(
+                {
+                    "error": "Collection cannot be delete because many products are associates with this"
+                },
+                status=status.HTTP_405_METHOD_NOT_ALLOWED,
+            )
         collection.delete()
         return Response(status=status.HTTP_204_NO_CONTENT)
